backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
azure-core==1.33.0
azure-cosmos==4.9.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distlib==0.3.9
distro==1.9.0
Django==4.2.20
django-multiselectfield==0.1.13
django-multiupload==0.6.1
djangorestframework==3.16.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.9.0
openai==1.70.0
pillow==11.1.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
whitenoise==6.9.0
